+++ account-keeper-sdk/src/main/java/com/dwarfeng/acckeeper/sdk/util/Constraints.java
@@ -17,9 +17,13 @@ public final class Constraints {
      */
     public static final int LENGTH_REMARK = 100;
     /**
-     * 密码的长度（此处指加密过的密码的长度）。
+     * 密码的长度（此处指加密过的密码的长度）约束。
      */
     public static final int LENGTH_PASSWORD = 60;
+    /**
+     * 标签的长度约束。
+     */
+    public static final int LENGTH_LABEL = 20;
 
     private Constraints() {
         throw new IllegalStateException("禁止实例化");