+++ account-keeper-stack/src/main/java/com/dwarfeng/acckeeper/stack/bean/entity/Account.java
@@ -11,23 +11,30 @@ import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
  */
 public class Account implements Entity<StringIdKey> {
 
-    private static final long serialVersionUID = -7166808835407200393L;
+    private static final long serialVersionUID = -7041696490265336859L;
 
     private StringIdKey key;
     private String password;
     private boolean enabled;
     private String remark;
     private long serialVersion;
+    /**
+     * @since 1.3.1
+     */
+    private String displayName;
 
     public Account() {
     }
 
-    public Account(StringIdKey key, String password, boolean enabled, String remark, long serialVersion) {
+    public Account(
+            StringIdKey key, String password, boolean enabled, String remark, long serialVersion, String displayName
+    ) {
         this.key = key;
         this.password = password;
         this.enabled = enabled;
         this.remark = remark;
         this.serialVersion = serialVersion;
+        this.displayName = displayName;
     }
 
     @Override
@@ -72,6 +79,14 @@ public class Account implements Entity<StringIdKey> {
         this.serialVersion = serialVersion;
     }
 
+    public String getDisplayName() {
+        return displayName;
+    }
+
+    public void setDisplayName(String displayName) {
+        this.displayName = displayName;
+    }
+
     @Override
     public String toString() {
         return "Account{" +
@@ -80,6 +95,7 @@ public class Account implements Entity<StringIdKey> {
                 ", enabled=" + enabled +
                 ", remark='" + remark + '\'' +
                 ", serialVersion=" + serialVersion +
+                ", displayName='" + displayName + '\'' +
                 '}';
     }
 }